requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
pdfplumber==0.9.0
openai==0.27.8
Werkzeug==2.3.6
